Fighter Extract
This is my code that i run to iterate through the API, as it looks at the CSV file of fighter names. My issue is every time i run this i am getting this error message: "JSONDecodeError: Expecting value: line 1 column 1 (char 0)". 

based on what i looked up this seems to be an issue with the data it is taking in not being in JSON format. i have tried to debug in VS, but i have hit a wall. Please help with solving this issue.